todo unit tests for amends, and all of returns everything
