+++ app/controllers/amends/AmendDirectExportPlasticPackagingController.scala
@@ -24,7 +24,7 @@ import forms.AmendDirectExportPlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.AmendDirectExportPlasticPackagingPage
+import pages.amends.AmendDirectExportPlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
