+++ app/controllers/amends/AmendHumanMedicinePlasticPackagingController.scala
@@ -24,7 +24,7 @@ import forms.AmendHumanMedicinePlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.AmendHumanMedicinePlasticPackagingPage
+import pages.amends.AmendHumanMedicinePlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
