+++ app/controllers/amends/AmendImportedPlasticPackagingController.scala
@@ -24,7 +24,7 @@ import forms.AmendImportedPlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.AmendImportedPlasticPackagingPage
+import pages.amends.AmendImportedPlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
