+++ app/controllers/amends/AmendManufacturedPlasticPackagingController.scala
@@ -24,7 +24,7 @@ import forms.AmendManufacturedPlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.AmendManufacturedPlasticPackagingPage
+import pages.amends.AmendManufacturedPlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
