+++ app/controllers/amends/AmendRecycledPlasticPackagingController.scala
@@ -24,7 +24,7 @@ import forms.AmendRecycledPlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.AmendRecycledPlasticPackagingPage
+import pages.amends.AmendRecycledPlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
