+++ app/controllers/helpers/TaxReturnHelper.scala
@@ -21,10 +21,12 @@ import models.UserAnswers
 import models.returns.ReturnType.{AMEND, NEW, ReturnType}
 import models.returns._
 import pages._
+import pages.amends.{AmendDirectExportPlasticPackagingPage, AmendHumanMedicinePlasticPackagingPage, AmendImportedPlasticPackagingPage, AmendManufacturedPlasticPackagingPage, AmendRecycledPlasticPackagingPage}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.MessagesControllerComponents
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
+
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
