+++ app/pages/amends/AmendRecycledPlasticPackagingPage.scala
@@ -14,13 +14,16 @@
  * limitations under the License.
  */
 
-package pages
+package pages.amends
 
+import pages.QuestionPage
 import play.api.libs.json.JsPath
 
-case object AmendDirectExportPlasticPackagingPage extends QuestionPage[Int] {
+trait AmendJourney
+
+case object AmendRecycledPlasticPackagingPage extends AmendJourney with QuestionPage[Int] {
 
   override def path: JsPath = JsPath \ "amend" \ toString
 
-  override def toString: String = "amendDirectExportPlasticPackaging"
+  override def toString: String = "amendRecycledPlasticPackaging"
 }
