+++ app/viewmodels/checkAnswers/amends/AmendRecycledPlasticPackagingSummary.scala
@@ -18,7 +18,7 @@ package viewmodels.checkAnswers.amends
 
 import controllers.routes
 import models.{CheckMode, UserAnswers}
-import pages.AmendRecycledPlasticPackagingPage
+import pages.amends.AmendRecycledPlasticPackagingPage
 import play.api.i18n.Messages
 import uk.gov.hmrc.govukfrontend.views.viewmodels.summarylist.SummaryListRow
 import viewmodels.checkAnswers.SummaryViewModel
