+++ test/controllers/AmendAreYouSureControllerSpec.scala
@@ -25,7 +25,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.{any, refEq}
 import org.mockito.Mockito.{verify, when}
 import org.scalatestplus.mockito.MockitoSugar
-import pages.AmendAreYouSurePage
+import pages.amends.AmendAreYouSurePage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
