+++ test/controllers/AmendDirectExportPlasticPackagingControllerSpec.scala
@@ -24,7 +24,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.AmendDirectExportPlasticPackagingPage
+import pages.amends.AmendDirectExportPlasticPackagingPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
