+++ test/controllers/AmendRecycledPlasticPackagingControllerSpec.scala
@@ -24,7 +24,7 @@ import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.AmendRecycledPlasticPackagingPage
+import pages.amends.AmendRecycledPlasticPackagingPage
 import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
